+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Search tasks (eg. Export) work with Member,Contribute,Participant,Case...
 * Added source to retrieve the owner membership, when owner membership is not set (meaning it is already the primary) then it will return itself.
 * Added date filter to filter date with the PHP Date Format.
+* Added filtering on Contact (sub) type on the contact filter.
 
 # Version 1.3.0